spatial: cast categorical codes to int64 (pair-code int8 overflow)
The real MERFISH object has 13 cell types and no NaN labels: the bincount
ValueError was int8 overflow in cluster-pair codes (12*13 > 127), not
missing annotations. Regression test with 14 clusters added.

+++ src/module4_spatial_aging/spatial_analysis.py
@@ -106,7 +106,8 @@ def neighborhood_enrichment_z(
     cats, k = cat.categories, len(cat.categories)
     if (cat.codes < 0).any():
         raise ValueError(f"labels contain {int((cat.codes < 0).sum())} NaN values")
-    code = cat.codes
+    # int64: codes are int8/int16 for small k; k*k pair codes would overflow
+    code = cat.codes.astype(np.int64)
     coo = connectivities.tocoo()
 
     def pair_counts(c: np.ndarray) -> np.ndarray:
